# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/company/list-company-account/{companyId}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Company",
                "operationId": "f03c7d2371c289c7537be9d1ac992dc2",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 4
                                                    },
                                                    "companyId": {
                                                        "type": "number",
                                                        "example": 6
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Prof. Jamir Stanton Sr."
                                                    },
                                                    "nameRomaji": {
                                                        "type": "string",
                                                        "example": "Jamir Stanton"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "cfunk@example.com"
                                                    },
                                                    "sex": {
                                                        "type": "number",
                                                        "example": 0
                                                    },
                                                    "dateOfBirth": {
                                                        "type": "string",
                                                        "example": "2022-05-19"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+19732341225"
                                                    },
                                                    "roleId": {
                                                        "type": "number",
                                                        "example": 2
                                                    },
                                                    "position": {
                                                        "type": "string",
                                                        "example": "Chief Executive Officer"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "Yasmin Kuphal.png"
                                                    },
                                                    "status": {
                                                        "type": "number",
                                                        "example": 0
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad validation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "jwt_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token\\>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}